fovi/requirements.txt
einops
fastcore
hydra-core
ipython
matplotlib
numpy
omegaconf
pandas
Pillow
pygeodesic
scikit_learn
scipy
seaborn
scikit-image
timm
torch
torchmetrics
torchvision
tqdm
trimesh
wandb
numba
plum-dispatch
transformers
cmasher
ipywidgets
fvcore
accelerate